+++ verik-compiler/src/main/kotlin/io/verik/compiler/transform/pre/BinaryExpressionReducerStage.kt
@@ -46,6 +46,8 @@ object BinaryExpressionReducerStage : ProjectStage() {
             Core.Kt.Double.F_plus_Double
         referenceMap[ReducerEntry(Core.Kt.C_Double, Core.Kt.C_Int, KtBinaryOperatorKind.DIV)] =
             Core.Kt.Double.F_div_Int
+        referenceMap[ReducerEntry(Core.Kt.C_String, Core.Kt.C_String, KtBinaryOperatorKind.PLUS)] =
+            Core.Kt.String.F_plus_Any
         referenceMap[ReducerEntry(Core.Vk.C_Ubit, Core.Vk.C_Ubit, KtBinaryOperatorKind.PLUS)] =
             Core.Vk.Ubit.F_plus_Ubit
         referenceMap[ReducerEntry(Core.Vk.C_Ubit, Core.Vk.C_Sbit, KtBinaryOperatorKind.PLUS)] =
@@ -86,6 +88,7 @@ object BinaryExpressionReducerStage : ProjectStage() {
 
     private object BinaryExpressionReducerVisitor : TreeVisitor() {
 
+        // TODO generalize reduce for Core.Kt.String.F_plus_Any
         override fun visitKtBinaryExpression(binaryExpression: EKtBinaryExpression) {
             super.visitKtBinaryExpression(binaryExpression)
             val receiverDeclaration = binaryExpression.left.type.reference