+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support cover bins for `CoverPoint` and `CoverCross`.
 - Support `fill0`, `fill1`, `fillx`, and `fillz`.
 - Support `rangeTo` operator.
+- Support string concatenation.
 
 ## [0.1.14]
 ### Importer